% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inputs.R
\name{build_inputs}
\alias{build_inputs}
\title{Build Inputs from Script Metadata for Phuse Web Framework}
\usage{
build_inputs(fn = NULL)
}
\arguments{
\item{fn}{a filen name or URL pointing to script metadata file}
}
\value{
R shiny code for providing inputs to the script
}
\description{
Build R shiny code for Phuse Web Apps
}
\author{
Hanming Tu
}
