% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_inputs.R
\name{buiuld_inputs}
\alias{buiuld_inputs}
\alias{build_inputs}
\title{Build Inputs from Script Metadata for Phuse Web Framework}
\usage{
build_inputs(fn = NULL)
}
\arguments{
\item{fn}{a file name or URL pointing to script metadata file}
}
\value{
R shiny code for providing inputs to the script
}
\description{
Build R shiny code for Phuse Web Apps
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  r1 <- build_inputs(f1)
}
\author{
Hanming Tu
}
