% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_workdir.R
\name{chk_workdir}
\alias{chk_workdir}
\title{Verify work directory}
\usage{
chk_workdir(dir, top_dir = NULL, sub_dir = "myRepo")
}
\arguments{
\item{dir}{a work directory; default to '/Users/{user} for Mac; "c:/tmp" for Windows}

\item{top_dir}{a top or root directory; default to '/Users/{user} for Mac
or getwd for other OS}

\item{sub_dir}{a sub directory}
}
\value{
TRUE or FALSE
}
\description{
Verify if the dir is the work directory
}
\examples{
  f1 <- "/Users/htu/myRepo"
  r1 <- chk_workdir(f1)
}
\author{
Hanming Tu
}
