% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_github.R
\name{clone_github}
\alias{clone_github}
\title{Clone a GitHub repositoru}
\usage{

  clone_github(repo_url = "https://github.com/phuse-org/phuse-scripts.git",
  repo_dir = NULL, repo_name = NULL, upd_opt = NULL)
}
\arguments{
\item{repo_url}{a URL for a remote repository and default to
'https://github.com/phuse-org/phuse-scripts.git'}

\item{repo_dir}{a local directory to host the repository;
default to work_dir from crt_workdir if not specified}

\item{repo_name}{repo name; default to the repo name in repo_url.}

\item{upd_opt}{update option: Repo}
}
\value{
nothing.
}
\description{
Clone a specified GitHub.
}
\examples{
\dontrun{
  r1 <- clone_github()
}
}
\author{
Hanming Tu
}
