% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(r_dir, s_dir = NULL)
}
\arguments{
\item{r_dir}{root directory}

\item{s_dir}{sub directory}
}
\value{
directory name
}
\description{
create a directory
}
\author{
Hanming Tu
}
