% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvt_class2df.R
\name{cvt_class2df}
\alias{cvt_class2df}
\title{Convert a class to data fram}
\usage{
cvt_class2df(x, exc = "^__", condition = FALSE)
}
\arguments{
\item{x}{a class or list}

\item{exc}{exclude pattern}

\item{condition}{condition for excluding}
}
\description{
Convert class or list to a data frame
}
\examples{
  r1 <- Sys.getenv()
  r2 <- cvt_class2df(r1)
}
\author{
Hanming Tu
}
