% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvt_list2df.R
\name{cvt_list2df}
\alias{cvt_list2df}
\title{Convert list to data frame}
\usage{
cvt_list2df(a)
}
\arguments{
\item{a}{a list returned by read_yml or any list}
}
\value{
data frame
}
\description{
convert list to a data frame with the following structure:
  variable, level, type, value
}
\author{
Hanming Tu
}
