% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fns.R
\name{download_fns}
\alias{download_fns}
\title{Download files from a repository}
\usage{
download_fns(df, tgtDir = NULL,
  baseDir = "https://github.com/phuse-org/phuse-scripts/raw/master")
}
\arguments{
\item{df}{a data frame containing file names produced from extract_fns}

\item{tgtDir}{target directory for storing the files}

\item{baseDir}{base directory in the repository including the repo URL.
Default to "https://github.com/phuse-org/phuse-scripts/raw/master"}
}
\description{
download files defined in the input data frame.
}
\author{
Hanming Tu
}
