% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_script.R
\name{download_script}
\alias{download_script}
\title{Download files defined in script metadata}
\usage{
download_script(cfg, wkDir = "workdir", source_lib = TRUE)
}
\arguments{
\item{cfg}{a list containing script metadata}

\item{wkDir}{work directory where the files will be downloaded to}

\item{source_lib}{whether to source the library defined for the scrpt in the metadata}
}
\value{
target directory name
}
\description{
download scripts, data or any file defined in the script metadata.
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  f2 <- read_yml(f1)
  f3 <- download_script(f2)
}
\author{
Hanming Tu
}
