% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_script_files.R
\name{download_script_files}
\alias{download_script_files}
\title{Download files from a repository}
\usage{
download_script_files(fns, tgtDir,
  baseDir = "https://github.com/phuse-org/phuse-scripts/raw/master",
  scriptDir = "data/send/PDS/Xpt")
}
\arguments{
\item{fns}{a list containing file names}

\item{tgtDir}{target directory for storing the files}

\item{baseDir}{base directory in the repository including the repo URL.
Default to "https://github.com/phuse-org/phuse-scripts/raw/master"}

\item{scriptDir}{script directory in the repository}
}
\description{
download files defined in the input list from a repository.
}
\examples{
  fns <- c("dm.xpt","ex.xpt")
  dir <- "/Users/htu/myRepo/data"
  # a <- download_script_files(fns, dir)
}
\author{
Hanming Tu
}
