% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fns.R
\name{extract_fns}
\alias{extract_fns}
\title{Extract File Names from Script Metadata}
\usage{
extract_fns(lst)
}
\arguments{
\item{lst}{a list containing script metadata}
}
\value{
a data frame (subdir, filename) containing parsed file names
}
\description{
extract folders and file names from a list containing script metadata.
}
\author{
Hanming Tu
}
