% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inputs.R
\name{get_inputs}
\alias{get_inputs}
\title{Get Inputs from Input Sources}
\usage{
get_inputs(fn = NULL)
}
\arguments{
\item{fn}{a file name or URL pointing to script metadata file}
}
\value{
a list of input values provided for the script
}
\description{
Get inputs from interactive session (shiny webpage), command line or script metadata.
}
\author{
Hanming Tu
}
