% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{Check if a variable is na or null or space}
\usage{
is_empty(x)
}
\arguments{
\item{x}{a list or string}
}
\value{
true or false
}
\description{
check if string or list is empty (na, null or blank spaces).
}
\examples{
  is_empty(NULL);
  is_empty('');
  is_empty(NA);
}
\author{
Hanming Tu
}
