% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{Convert list to data frame}
\usage{
list2df(a)
}
\arguments{
\item{a}{the 1st list}
}
\value{
data frame
}
\description{
convert list to a data frame with the following structure:
  variable, level, type, value
}
\author{
Hanming Tu
}
