% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\title{Resolve absolute path}
\usage{
resolve(dir, relpath)
}
\arguments{
\item{dir}{directory}

\item{relpath}{relative path}
}
\description{
Resolve absolute directory
}
\author{
Hanming Tu
}
