% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run example}
\usage{
run_example(example = NA, pkg = "phuse", port = NULL,
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"), display.mode = c("auto",
  "normal", "showcase"))
}
\arguments{
\item{example}{Example name}

\item{pkg}{package name}

\item{port}{Port number}

\item{launch.browser}{define the browser- shiny.launch.browser}

\item{host}{define the host or ip address}

\item{display.mode}{modes are auto, normal or showcase}
}
\description{
run examples stored in the example folder.
}
\examples{
  library(phuse)
  # I have to comment it out so it can pass "R CMD check --as-cran"
  # run_example("02_display")
}
\author{
Hanming Tu
}
