% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_github.R
\name{search_github}
\alias{search_github}
\title{Search GitHub and build a script index data frame}
\usage{
search_github(filename = "*.yml",
  rep_url = "https://github.com/phuse-org/phuse-scripts",
  rep_dir = "tree/master", out_type = "fn")
}
\arguments{
\item{filename}{file names to be searched; default to *.yml.}

\item{rep_url}{a URL for a remote repository and default to
'https://github.com/phuse-org/phuse-scripts.git'}

\item{rep_dir}{rep dir for file name}

\item{out_type}{output type; default to 'fn' - just file names.}
}
\value{
a list of YML files
}
\description{
Use GitHub search API to search for YML files in phuse-scripts
  repository and output the list.
}
\examples{
\dontrun{
  r1 <- search_github('*.yml')
}
}
\author{
Hanming Tu
}
