% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_app.R
\name{start_app}
\alias{start_app}
\title{Start Phuse Web Application}
\usage{
start_app(app_name = NULL, n = 2, pkg = "phuse", pt = NULL,
  lb = getOption("shiny.launch.browser", interactive()),
  ht = getOption("shiny.host", "127.0.0.1"), dm = "normal",
  msg_lvl = NULL, loc = "local")
}
\arguments{
\item{app_name}{app or script name}

\item{n}{Example number}

\item{pkg}{package name}

\item{pt}{Port number}

\item{lb}{define the browser- shiny.launch.browser}

\item{ht}{define the host or ip address}

\item{dm}{display modes are auto, normal or showcase}

\item{msg_lvl}{message level}

\item{loc}{location of the scirpt: local|github; default to 'local'}
}
\description{
start phuse web appllication framework. This includes all the
  functions in start_phuse plus starting standalone application by name.
}
\examples{
\dontrun{
  library(phuse)
  start_appe()  # default to "02_display"
  start_app(1)  # start "01_html"
}
}
\author{
Hanming Tu
}
