% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_phuse.R
\name{start_phuse}
\alias{start_phuse}
\title{Start Phuse Web Application}
\usage{
start_phuse(n = 2, pkg = "phuse", pt = NULL,
  lb = getOption("shiny.launch.browser", interactive()),
  ht = getOption("shiny.host", "127.0.0.1"), dm = "normal",
  msg_lvl = NULL)
}
\arguments{
\item{n}{Example number}

\item{pkg}{package name}

\item{pt}{Port number}

\item{lb}{define the browser- shiny.launch.browser}

\item{ht}{define the host or ip address}

\item{dm}{display modes are auto, normal or showcase}

\item{msg_lvl}{message level}
}
\description{
start phuse web appllication framework.
}
\examples{
\dontrun{
  library(phuse)
  start_phusee()  # default to "02_display"
  start_phuse(1)  # start "01_html"
}
}
\author{
Hanming Tu
}
