% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url.exists}
\alias{url.exists}
\title{Check URL based on httr package}
\usage{
url.exists(url = "https://github.com/phuse-org/phuse-scripts.git",
  show = FALSE)
}
\arguments{
\item{url}{a URL for a remote repository and default to
'https://github.com/phuse-org/phuse-scripts.git'}

\item{show}{bolean variable; default to FALSE}
}
\value{
TRUE or FALSE
}
\description{
Check if URL exists.
}
\examples{
  url.exists('https://github.com/phuse-org/phuse-scripts.git')
}
\author{
Hanming Tu
}
