% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_workdir.R
\name{crt_workdir}
\alias{crt_workdir}
\title{Create work directory}
\usage{
crt_workdir(top_dir = NULL, sub_dir = "myRepo", to_crt_dir = TRUE)
}
\arguments{
\item{top_dir}{a top or root directory; default to '/Users/{user} for Mac
or getwd for other OS}

\item{sub_dir}{a sub directory}

\item{to_crt_dir}{whether to create the dir; default to TRUE.
If FALSE, just return the dir name}
}
\value{
the created directory
}
\description{
define and create a work directory.
}
\examples{
  f1 <- "/Users/htu"
  r1 <- crt_workdir(f1)
}
\author{
Hanming Tu
}
