% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvt_list2df.R
\name{cvt_list2df}
\alias{cvt_list2df}
\title{Convert list to data frame}
\usage{
cvt_list2df(a)
}
\arguments{
\item{a}{a list returned by read_yml or any list}
}
\value{
data frame
}
\description{
convert list to a data frame with the following structure:
  variable, level, type, value
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  r1 <- read_yml(f1)
  r2 <- cvt_list2df(r1)
}
\author{
Hanming Tu
}
