% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fns.R
\name{extract_fns}
\alias{extract_fns}
\title{Extract File Names from Script Metadata}
\usage{
extract_fns(lst)
}
\arguments{
\item{lst}{a list containing script metadata}
}
\value{
a data frame (subdir, filename) containing parsed file names
}
\description{
extract folders and file names from a list containing script metadata.
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  f2 <- read_yml(f1)
  f3 <- extract_fns(f2)
}
\author{
Hanming Tu
}
