% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yml_inputs.R
\name{get_yml_inputs}
\alias{get_yml_inputs}
\title{Get Inputs from Script Metadata File}
\usage{
get_yml_inputs(fn = NULL)
}
\arguments{
\item{fn}{a file name or URL pointing to script metadata file}
}
\value{
a list of input values provided for the script
}
\description{
Get inputs from script metadata.
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  r1 <- get_inputs(f1)
}
\author{
Hanming Tu
}
