% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cfg.R
\name{init_cfg}
\alias{init_cfg}
\title{Initialize configuration for phuse}
\usage{
init_cfg(cfg)
}
\arguments{
\item{cfg}{a list containing script metadata information}
}
\value{
a list containing the merged configuration
}
\description{
read script metadata file in the repository and merged it
  with a local script metadata file if it exists.
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  f2 <- read_yml(f1)
  r1 <- init_cfg(f2)
}
\author{
Hanming Tu
}
