% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yml.R
\name{read_yml}
\alias{read_yml}
\title{Read YML file into a list}
\usage{
read_yml(fn)
}
\arguments{
\item{fn}{a URL or file name containing script metadata}
}
\value{
a list containing the parsed metadata tags
}
\description{
read script metadata file in the repository or in a local folder.
}
\examples{
  a <- "https://github.com/phuse-org/phuse-scripts/raw/master"
  b <- "development/R/scripts"
  c <- "Draw_Dist2_R.yml"
  f1 <- paste(a,b,c, sep = '/')
  r1 <- get_inputs(f1)
}
\author{
Hanming Tu
}
