% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_phuse.R
\name{start_phuse}
\alias{start_phuse}
\title{Start Phuse Web Application}
\usage{
start_phuse(n = 2)
}
\arguments{
\item{n}{Example number}
}
\description{
start phuse web appllication framework.
}
\examples{
  library(phuse)
  # comment out the interactive sessions
  # start_phusee()  # default to "02_display"
  # start_phuse(1)  # start "01_html"
}
\author{
Hanming Tu
}
